requirements.txt
requests==2.26.0
selenium==4.1.0
pytest==6.2.5
jsonschema==4.2.1
pytest-xdist==2.5.0
allure-pytest==2.9.45